RENAME reducers folder
